+++ lib/addStyles.js
@@ -3,7 +3,7 @@
 	Author Tobias Koppers @sokra
 */
 
-var exportedStyles = require("../export");
+var exportedStyles = require("../exports");
 
 var stylesInDom = {};
